requirements.txt
pyaml
torch == 2.1.1
langchain == 0.0.339
ctransformers == 0.2.27